ddi/llm/utils.py
# ddi/llm/utils.py

from __future__ import annotations
import logging
import json
import re
from typing import Tuple, List, Optional, Type, Any, Dict

from pydantic import BaseModel, ValidationError, TypeAdapter
from google.genai import types

from ddi.llm.providers import generate_response
from ddi.schemas.item_schemas import RefinementPatch
from ddi.schemas.models import Item
from ddi.prompts import load_prompt
from ddi.pipelines.utils.parsers import extract_json_block, build_prompt_messages, parse_payload

logger = logging.getLogger(__name__)


def _preprocess_multiline_blocks(raw_text: str) -> str:
    """
    Busca bloques de texto multilínea delimitados y los convierte en
    strings JSON escapados válidos.
    """
    multiline_pattern = re.compile(
        r'"(.*?)"\s*:\s*"<<<TEXTO_MULTILINEA\n(.*?)\nTEXTO_MULTILINEA>>>"',
        re.DOTALL
    )

    def escape_and_replace(match):
        field_name = match.group(1)
        content_to_escape = match.group(2)
        escaped_content = json.dumps(content_to_escape)
        return f'"{field_name}": {escaped_content}'

    return multiline_pattern.sub(escape_and_replace, raw_text)

async def call_llm_and_parse_json_result(
    prompt_name: str,
    user_input_content: str,
    stage_name: str,
    item: Item,
    ctx: Dict[str, Any],
    expected_schema: Optional[Type[BaseModel]] = None,
    **kwargs,
) -> Tuple[Optional[BaseModel | str], Optional[List[RefinementPatch]], int]:
    """
    Llama a un LLM, limpia la respuesta y valida el resultado contra un esquema Pydantic.
    """
    total_tokens_used = 0
    response_text = ""
    llm_response = None
    try:
        prompt_data = load_prompt(prompt_name)
        system_template = prompt_data.get("system_message", "") if isinstance(prompt_data, dict) else ""
        user_prompt_template = prompt_data.get("content", "") if isinstance(prompt_data, dict) else prompt_data

        if not user_prompt_template:
            raise ValueError(f"No se pudo cargar una plantilla de prompt válida desde '{prompt_name}'.")

        messages = build_prompt_messages(system_template, user_prompt_template, user_input_content)

        llm_response = await generate_response(messages=messages, **kwargs)
        tokens_used = llm_response.usage.get("total", 0)
        total_tokens_used += tokens_used
        item.token_usage += tokens_used

        if not llm_response.success:
            error_msg = llm_response.error_message or "Error desconocido del proveedor LLM."
            error = RefinementPatch(code="E905_LLM_CALL_FAILED", field_path="llm_response", description=error_msg)
            return None, [error], total_tokens_used

        response_text = llm_response.text
        if not response_text:
            error = RefinementPatch(code="E904_LLM_NO_RESPONSE", field_path="llm_response", description="El LLM no devolvió contenido.")
            return None, [error], total_tokens_used

        clean_text = extract_json_block(response_text)
        processed_text = _preprocess_multiline_blocks(clean_text)

        if expected_schema is None:
            return processed_text, None, total_tokens_used

        payload_data = json.loads(processed_text, strict=False)
        validated_obj = TypeAdapter(expected_schema).validate_python(payload_data)

        return validated_obj, None, total_tokens_used

    except (json.JSONDecodeError, ValidationError) as e:
        raw_response_text = llm_response.text if llm_response and hasattr(llm_response, 'text') else "No se pudo obtener la respuesta de texto."
        logger.error(
            f"[{stage_name}] Item {item.temp_id}: Falló la validación/parseo. Respuesta cruda: \n{raw_response_text}"
        )
        error_msg = f"Error parseando/validando respuesta: {e}. Respuesta: {response_text[:500]}..."
        error = RefinementPatch(code="E904_LLM_RESPONSE_FORMAT_ERROR", field_path="llm_response", description=error_msg)
        return None, [error], total_tokens_used
    except Exception as e:
        error_msg = f"Error inesperado en la utilidad LLM: {e}"
        logger.error(f"[{stage_name}] Item {item.temp_id}: {error_msg}", exc_info=True)
        error = RefinementPatch(code="E999_UNEXPECTED_ERROR", field_path="llm_call", description=error_msg)
        return None, [error], total_tokens_used

async def call_llm_with_tools(
    prompt_name: str,
    user_input_content: str,
    stage_name: str,
    item: Item,
    ctx: Dict[str, Any],
    expected_schema: Type[BaseModel],
    tools: Optional[List[Any]] = None,
    **kwargs,
) -> Tuple[Optional[BaseModel], Optional[List[RefinementPatch]], int]:
    """
    Maneja un flujo de agente con herramientas y usa TypeAdapter para la validación final.
    """
    total_tokens_used = 0
    # ... (lógica completa de Sigie para manejo de herramientas) ...
    # Por brevedad, se omite la lógica interna, pero se mantendría la de Sigie.
    # El punto clave es que esta función ahora llama a `generate_response` sin crear dependencias circulares.
    pass
